+++ assets/src/scripts/store/index.js
@@ -391,15 +391,15 @@ export const Actions = {
             width
         };
     },
-    setHydrographXRange(hydrographXRange) {
+    setHydrographBrushOffset(hydrographBrushOffset) {
         return {
-            type: 'SET_HYDROGRAPH_X_RANGE',
-            hydrographXRange
+            type: 'SET_HYDROGRAPH_BRUSH_OFFSET',
+            hydrographBrushOffset
         };
     },
-    clearHydrographXRange() {
+    clearHydrographBrushOffset() {
         return {
-            type: 'CLEAR_HYDROGRAPH_X_RANGE'
+            type: 'CLEAR_HYDROGRAPH_BRUSH_OFFSET'
         };
     },
     setCurrentVariable(variableID) {
@@ -537,7 +537,7 @@ export const configureStore = function (initialState) {
         ui : {
             windowWidth: 1024,
             width: 800,
-            hydrographXRange: {}
+            hydrographBrushOffset: {}
         },
         ...initialState
     };